back-end/requirements.txt
Flask
Flask-SQLAlchemy
mysqlclient
marshmallow
PyJWT
flask-cors